+++ src/vivarium/framework/configuration.py
@@ -8,43 +8,49 @@ representations of :term:`model specifications <Model Specification>` and
 :term:`configurations <Configuration>`.
 
 """
-
-import os.path
+from pathlib import Path
+from typing import Union, Dict
 
 import yaml
 
-from vivarium.exceptions import VivariumError
-from vivarium.config_tree import ConfigTree
-
-from .plugins import DEFAULT_PLUGINS
+from vivarium.config_tree import ConfigTree, ConfigurationError
+from vivarium.framework.plugins import DEFAULT_PLUGINS
 
 
-class ConfigurationError(VivariumError):
-    """Error raised when invalid configuration is received."""
-    pass
+def build_model_specification(model_specification: Union[str, Path, ConfigTree] = None,
+                              component_configuration: Union[Dict, ConfigTree] = None,
+                              configuration: Union[Dict, ConfigTree] = None,
+                              plugin_configuration: Union[Dict, ConfigTree] = None) -> ConfigTree:
+    if isinstance(model_specification, (str, Path)):
+        validate_model_specification_file(model_specification)
+
+    output_spec = _get_default_specification()
+    output_spec.update(model_specification, layer='model_override', source=str(model_specification))
+    output_spec.components.update(component_configuration, layer='override', source='user_supplied_args')
+    output_spec.configuration.update(configuration, layer='override', source='user_supplied_args')
+    output_spec.plugins.update(plugin_configuration, layer='override', source='user_supplied_args')
+
+    return output_spec
 
 
-def build_model_specification(model_specification_file_path: str) -> ConfigTree:
-    model_specification_file_path = validate_model_specification_file(model_specification_file_path)
-
-    model_specification = _get_default_specification()
-    model_specification.update(model_specification_file_path, layer='model_override')
-
-    return model_specification
-
-
-def validate_model_specification_file(file_path: str) -> str:
+def validate_model_specification_file(file_path: Union[str, Path]) -> None:
     """Ensures the provided file is a yaml file"""
-    if not os.path.isfile(file_path):
+    file_path = Path(file_path)
+    if not file_path.exists():
         raise ConfigurationError('If you provide a model specification file, it must be a file. '
-                                 f'You provided {file_path}')
+                                 f'You provided {str(file_path)}', value_name=None)
 
-    extension = file_path.split('.')[-1]
-    if extension not in ['yaml', 'yml']:
-        raise ConfigurationError(f'Model specification files must be in a yaml format. You provided {extension}')
+    if file_path.suffix not in ['.yaml', '.yml']:
+        raise ConfigurationError(f'Model specification files must be in a yaml format. You provided {file_path.suffix}',
+                                 value_name=None)
     # Attempt to load
-    yaml.full_load(file_path)
-    return file_path
+    with file_path.open() as f:
+        raw_spec = yaml.full_load(f)
+    top_keys = set(raw_spec.keys())
+    valid_keys = {'plugins', 'components', 'configuration'}
+    if not top_keys <= valid_keys:
+        raise ConfigurationError(f'Model specification contains additional top level '
+                                 f'keys {valid_keys.difference(top_keys)}.', value_name=None)
 
 
 def build_simulation_configuration() -> ConfigTree:
@@ -57,10 +63,11 @@ def _get_default_specification():
 
     model_specification = ConfigTree(layers=default_config_layers)
     model_specification.update(DEFAULT_PLUGINS, **default_metadata)
-    model_specification.update({'components': None})
+    model_specification.update({'components': {},
+                                'configuration': {}})
 
-    user_config_path = os.path.expanduser('~/vivarium.yaml')
-    if os.path.exists(user_config_path):
-        model_specification.configuration.update(user_config_path, layer='component_configs')
+    user_config_path = Path('~/vivarium.yaml').expanduser()
+    if user_config_path.exists():
+        model_specification.configuration.update(user_config_path, layer='base')
 
     return model_specification