+++ src/vivarium/interface/cli.py
@@ -29,21 +29,23 @@ simulations from the command line.
    :show-nested:
 
 """
-import os
-import shutil
-import cProfile
-import pstats
 from bdb import BdbQuit
+import cProfile
 from pathlib import Path
+import pstats
+from time import time
+import shutil
 
 import click
+import pandas as pd
 
-import vivarium
+from vivarium.examples import disease_model
 from vivarium.framework.engine import run_simulation
-from .utilities import verify_yaml
+from vivarium.framework.utilities import handle_exceptions
+from .utilities import get_output_root
 
 import logging
-logging.getLogger(__name__)
+_log = logging.getLogger(__name__)
 
 
 @click.group()
@@ -58,16 +60,15 @@ def simulate():
 
 
 @simulate.command()
-@click.argument('model_specification', callback=verify_yaml,
+@click.argument('model_specification',
                 type=click.Path(exists=True, dir_okay=False, resolve_path=True))
 @click.option('--results_directory', '-o', type=click.Path(resolve_path=True),
-              default=os.path.expanduser('~/vivarium_results/'),
+              default=Path('~/vivarium_results/').expanduser(),
               help='The directory to write results to. A folder will be created '
                    'in this directory with the same name as the configuration file.')
 @click.option('--verbose', '-v', is_flag=True, help='Report each time step.')
-@click.option('--log', type=click.Path(dir_okay=False, resolve_path=True), help='The path to write a log file to.')
 @click.option('--pdb', 'with_debugger', is_flag=True, help='Drop into python debugger if an error occurs.')
-def run(model_specification, results_directory, verbose, log, with_debugger):
+def run(model_specification, results_directory, verbose, with_debugger):
     """Run a simulation from the command line.
 
     The simulation itself is defined by the given MODEL_SPECIFICATION yaml file.
@@ -76,22 +77,22 @@ def run(model_specification, results_directory, verbose, log, with_debugger):
     is provided, a subdirectory will be created with the same name as the
     MODEL_SPECIFICATION if one does not exist. Results will be written to a
     further subdirectory named after the start time of the simulation run."""
-    log_level = logging.DEBUG if verbose else logging.ERROR
-    logging.basicConfig(filename=log, level=log_level)
+    start = time()
+    results_root = get_output_root(results_directory, model_specification)
+    results_root.mkdir(parents=True, exist_ok=False)
 
-    try:
-        run_simulation(model_specification, results_directory)
-    except (BdbQuit, KeyboardInterrupt):
-        raise
-    except Exception as e:
-        if with_debugger:
-            import pdb
-            import traceback
-            traceback.print_exc()
-            pdb.post_mortem()
-        else:
-            logging.exception("Uncaught exception {}".format(e))
-            raise
+    log_level = logging.DEBUG if verbose else logging.ERROR
+    logging.basicConfig(filename=results_root / 'simulation.log', level=log_level)
+    shutil.copy(model_specification, results_root / 'model_specification.yaml')
+
+    main = handle_exceptions(run_simulation, _log, with_debugger)
+    override_configuration = {'output_data': {'results_directory': str(results_root)}}
+    finished_sim = main(model_specification, configuration=override_configuration)
+
+    idx = pd.Index([finished_sim.configuration.randomness.random_seed], name='random_seed')
+    metrics = pd.DataFrame(finished_sim.report(), index=idx)
+    metrics['simulation_run_time'] = time() - start
+    metrics.to_hdf(results_root / 'output.hdf', key='data')
 
 
 @simulate.command()
@@ -100,32 +101,29 @@ def test():
     provided in the examples directory.
     """
     logging.basicConfig(level=logging.DEBUG)
-    model_specification = Path(vivarium.__file__).resolve().parent / 'examples' / 'disease_model' / 'disease_model.yaml'
-    results_directory = Path('~/vivarium_results').expanduser()
+    model_specification = disease_model.get_model_specification_path()
 
     try:
-        run_simulation(str(model_specification), str(results_directory))
+        run_simulation(model_specification)
         click.echo()
         click.secho("Installation test successful!", fg='green')
     except (BdbQuit, KeyboardInterrupt):
         raise
     except Exception as e:
-        logging.exception("Uncaught exception {}".format(e))
+        _log.exception("Uncaught exception {}".format(e))
         raise
-    finally:
-        shutil.rmtree(results_directory, ignore_errors=True)
 
 
 @simulate.command()
-@click.argument('model_specification', callback=verify_yaml,
+@click.argument('model_specification',
                 type=click.Path(exists=True, dir_okay=False, resolve_path=True))
 @click.option('--results_directory', '-o', type=click.Path(resolve_path=True),
-              default=os.path.expanduser('~/vivarium_results/'),
+              default=Path('~/vivarium_results/').expanduser(),
               help='The directory to write results to. A folder will be created '
                    'in this directory with the same name as the configuration file.')
 @click.option('--process/--no-process', default=False,
-               help=('Automatically process the profile to human readable format  with pstats, '
-                     'sorted by cumulative runtime, and dump to a file'))
+              help=('Automatically process the profile to human readable format  with pstats, '
+                    'sorted by cumulative runtime, and dump to a file'))
 def profile(model_specification, results_directory, process):
     """Run a simulation based on the provided MODEL_SPECIFICATION and profile
     the run.
@@ -134,12 +132,12 @@ def profile(model_specification, results_directory, process):
     results_directory = Path(results_directory)
 
     out_stats_file = results_directory / f'{model_specification.name}'.replace('yaml', 'stats')
-    command = f'run_simulation("{model_specification}", "{results_directory}")'
+    command = f'run_simulation("{model_specification}")'
     cProfile.runctx(command, globals=globals(), locals=locals(), filename=out_stats_file)
 
     if process:
         out_txt_file = results_directory / (out_stats_file.name + '.txt')
-        with open(out_txt_file, 'w') as f:
+        with out_txt_file.open('w') as f:
             p = pstats.Stats(str(out_stats_file), stream=f)
             p.sort_stats('cumulative')
             p.print_stats()