+++ tests/framework/test_state_machine.py
@@ -1,10 +1,9 @@
 import pandas as pd
 import numpy as np
 
+from vivarium import InteractiveContext
 from vivarium.framework.randomness import choice
 from vivarium.framework.state_machine import Machine, State, Transition
-
-from vivarium.interface.interactive import setup_simulation
 
 
 def _population_fixture(column, initial_value):
@@ -46,14 +45,14 @@ def test_transition():
     start_state.add_transition(done_state)
     machine = Machine('state', states=[start_state, done_state])
 
-    simulation = setup_simulation([machine, _population_fixture('state', 'start')])
-    event_time = simulation.clock.time + simulation.clock.step_size
+    simulation = InteractiveContext(components=[machine, _population_fixture('state', 'start')])
+    event_time = simulation._clock.time + simulation._clock.step_size
     machine.transition(simulation.get_population().index, event_time)
     assert np.all(simulation.get_population().state == 'done')
 
 
 def test_choice(base_config):
-    base_config.population.update({'population_size': 10000})
+    base_config.update({'population': {'population_size': 10000}})
     a_state = State('a')
     b_state = State('b')
     start_state = State('start')
@@ -61,15 +60,16 @@ def test_choice(base_config):
     start_state.add_transition(b_state, probability_func=lambda agents: np.full(len(agents), 0.5))
     machine = Machine('state', states=[start_state, a_state, b_state])
 
-    simulation = setup_simulation([machine, _population_fixture('state', 'start')], base_config)
-    event_time = simulation.clock.time + simulation.clock.step_size
+    simulation = InteractiveContext(components=[machine, _population_fixture('state', 'start')],
+                                    configuration=base_config)
+    event_time = simulation._clock.time + simulation._clock.step_size
     machine.transition(simulation.get_population().index, event_time)
     a_count = (simulation.get_population().state == 'a').sum()
     assert round(a_count/len(simulation.get_population()), 1) == 0.5
 
 
 def test_null_transition(base_config):
-    base_config.population.update({'population_size': 10000})
+    base_config.update({'population': {'population_size': 10000}})
     a_state = State('a')
     start_state = State('start')
     start_state.add_transition(a_state, probability_func=lambda agents: np.full(len(agents), 0.5))
@@ -77,15 +77,16 @@ def test_null_transition(base_config):
 
     machine = Machine('state', states=[start_state, a_state])
 
-    simulation = setup_simulation([machine, _population_fixture('state', 'start')], base_config)
-    event_time = simulation.clock.time + simulation.clock.step_size
+    simulation = InteractiveContext(components=[machine, _population_fixture('state', 'start')],
+                                    configuration=base_config)
+    event_time = simulation._clock.time + simulation._clock.step_size
     machine.transition(simulation.get_population().index, event_time)
     a_count = (simulation.get_population().state == 'a').sum()
     assert round(a_count/len(simulation.get_population()), 1) == 0.5
 
 
 def test_no_null_transition(base_config):
-    base_config.population.update({'population_size': 10000})
+    base_config.update({'population': {'population_size': 10000}})
     a_state = State('a')
     b_state = State('b')
     start_state = State('start')
@@ -96,8 +97,9 @@ def test_no_null_transition(base_config):
     machine = Machine('state')
     machine.states.extend([start_state, a_state, b_state])
 
-    simulation = setup_simulation([machine, _population_fixture('state', 'start')], base_config)
-    event_time = simulation.clock.time + simulation.clock.step_size
+    simulation = InteractiveContext(components=[machine, _population_fixture('state', 'start')],
+                                    configuration=base_config)
+    event_time = simulation._clock.time + simulation._clock.step_size
     machine.transition(simulation.get_population().index, event_time)
     a_count = (simulation.get_population().state == 'a').sum()
     assert round(a_count/len(simulation.get_population()), 1) == 0.5
@@ -125,8 +127,9 @@ def test_side_effects():
 
     machine = Machine('state', states=[start_state, done_state])
 
-    simulation = setup_simulation([machine, _population_fixture('state', 'start'), _population_fixture('count', 0)])
-    event_time = simulation.clock.time + simulation.clock.step_size
+    simulation = InteractiveContext(components=[machine, _population_fixture('state', 'start'),
+                                                _population_fixture('count', 0)])
+    event_time = simulation._clock.time + simulation._clock.step_size
     machine.transition(simulation.get_population().index, event_time)
     assert np.all(simulation.get_population()['count'] == 1)
     machine.transition(simulation.get_population().index, event_time)